+++ bigquery/src/main/scala/org/datatools/bigdatatypes/bigquery/BigQueryDefinitions.scala
@@ -3,7 +3,6 @@ package org.datatools.bigdatatypes.bigquery
 import com.google.cloud.bigquery.{Schema, StandardTableDefinition, TimePartitioning}
 import org.datatools.bigdatatypes.bigquery.JavaConverters.toJava
 
-
 private[bigquery] object BigQueryDefinitions {
 
   /** Given a column name created a TimePartition object */
@@ -19,47 +18,62 @@ private[bigquery] object BigQueryDefinitions {
     addPartitionToBuilder(builder, partitionColumn).build()
   }
 
-
   /** Given a builder and a column name, add a Time Partitioning column to the builder
-   *
-   * @param builder         for BigQuery tables
-   * @param partitionColumn name of the column to be used as Time Partitioning column
-   */
+    *
+    * @param builder         for BigQuery tables
+    * @param partitionColumn name of the column to be used as Time Partitioning column
+    */
   def addPartitionToBuilder(builder: StandardTableDefinition.Builder,
                             partitionColumn: Option[String]
-                           ): StandardTableDefinition.Builder =
+  ): StandardTableDefinition.Builder =
     partitionColumn.map(v => builder.setTimePartitioning(generateTimePartitionColumn(v))).getOrElse(builder)
 
   /** Generates a BigQuery Table Schema given a type A
-   */
+    */
   def generateSchema[A: SqlTypeToBigQuery]: Schema = Schema.of(toJava(SqlTypeToBigQuery[A].bigQueryFields))
 
   def generateSchema[A: SqlTypeToBigQuery, B: SqlTypeToBigQuery]: Schema =
     Schema.of(toJava(SqlTypeToBigQuery[A].bigQueryFields ++ SqlTypeToBigQuery[B].bigQueryFields))
 
   def generateSchema[A: SqlTypeToBigQuery, B: SqlTypeToBigQuery, C: SqlTypeToBigQuery]: Schema =
-    Schema.of(toJava(SqlTypeToBigQuery[A].bigQueryFields ++ SqlTypeToBigQuery[B].bigQueryFields ++ SqlTypeToBigQuery[C].bigQueryFields))
+    Schema.of(
+      toJava(
+        SqlTypeToBigQuery[A].bigQueryFields ++ SqlTypeToBigQuery[B].bigQueryFields ++ SqlTypeToBigQuery[
+          C
+        ].bigQueryFields
+      )
+    )
 
   def generateSchema[A: SqlTypeToBigQuery, B: SqlTypeToBigQuery, C: SqlTypeToBigQuery, D: SqlTypeToBigQuery]: Schema =
     Schema.of(
-      toJava(SqlTypeToBigQuery[A].bigQueryFields ++
-        SqlTypeToBigQuery[B].bigQueryFields ++
-        SqlTypeToBigQuery[C].bigQueryFields ++
-        SqlTypeToBigQuery[D].bigQueryFields)
+      toJava(
+        SqlTypeToBigQuery[A].bigQueryFields ++
+          SqlTypeToBigQuery[B].bigQueryFields ++
+          SqlTypeToBigQuery[C].bigQueryFields ++
+          SqlTypeToBigQuery[D].bigQueryFields
+      )
     )
 
-  def generateSchema[A: SqlTypeToBigQuery, B: SqlTypeToBigQuery, C: SqlTypeToBigQuery, D: SqlTypeToBigQuery, E: SqlTypeToBigQuery]: Schema =
+  def generateSchema[A: SqlTypeToBigQuery,
+                     B: SqlTypeToBigQuery,
+                     C: SqlTypeToBigQuery,
+                     D: SqlTypeToBigQuery,
+                     E: SqlTypeToBigQuery
+  ]: Schema =
     Schema.of(
-      toJava(SqlTypeToBigQuery[A].bigQueryFields ++
-        SqlTypeToBigQuery[B].bigQueryFields ++
-        SqlTypeToBigQuery[C].bigQueryFields ++
-        SqlTypeToBigQuery[D].bigQueryFields ++
-        SqlTypeToBigQuery[E].bigQueryFields)
+      toJava(
+        SqlTypeToBigQuery[A].bigQueryFields ++
+          SqlTypeToBigQuery[B].bigQueryFields ++
+          SqlTypeToBigQuery[C].bigQueryFields ++
+          SqlTypeToBigQuery[D].bigQueryFields ++
+          SqlTypeToBigQuery[E].bigQueryFields
+      )
     )
 
   /** For BigQueryTypesInstance
     * Multiples instances not supported for now
     */
-  def generateSchema[A: SqlInstanceToBigQuery](value: A): Schema = Schema.of(toJava(SqlInstanceToBigQuery[A].bigQueryFields(value)))
+  def generateSchema[A: SqlInstanceToBigQuery](value: A): Schema =
+    Schema.of(toJava(SqlInstanceToBigQuery[A].bigQueryFields(value)))
 
 }