+++ bigquery/src/test/scala/org/datatools/bigdatatypes/bigquery/SqlTypeToBigQueryInstanceSpec.scala
@@ -1,11 +1,11 @@
 package org.datatools.bigdatatypes.bigquery
 
 import com.google.cloud.bigquery.Field
-import org.datatools.bigdatatypes.TestTypes._
+import org.datatools.bigdatatypes.TestTypes.*
 import org.datatools.bigdatatypes.basictypes.SqlType
 import org.datatools.bigdatatypes.bigquery.SqlInstanceToBigQuery.InstanceSyntax
 import org.datatools.bigdatatypes.conversions.SqlTypeConversion
-import org.datatools.bigdatatypes.conversions.SqlTypeConversion._
+import org.datatools.bigdatatypes.conversions.SqlTypeConversion.*
 import org.datatools.bigdatatypes.formats.Formats.implicitDefaultFormats
 import org.datatools.bigdatatypes.{BigQueryTestTypes, UnitSpec}
 
@@ -19,7 +19,7 @@ class SqlTypeToBigQueryInstanceSpec extends UnitSpec {
 
   "SqlType Instance" should "have bigQueryFields method" in {
     val sql: SqlType = SqlTypeConversion[BasicTypes].getType
-    sql.bigQueryFields shouldBe BigQueryTestTypes.basicTypes
+    sql.asBigQuery shouldBe BigQueryTestTypes.basicTypes
   }
 
   "basic case class" should "be converted into BigQueryFields" in {