+++ core/src/main/scala_3/org/datatools/bigdatatypes/basictypes/SqlType.scala
@@ -1,7 +1,7 @@
 package org.datatools.bigdatatypes.basictypes
 
 import org.datatools.bigdatatypes.basictypes.SqlType.{SqlBool, SqlDate, SqlDecimal, SqlDouble, SqlFloat, SqlInt, SqlLong, SqlString, SqlStruct, SqlTimestamp}
-import org.datatools.bigdatatypes.basictypes.SqlTypeMode._
+import org.datatools.bigdatatypes.basictypes.SqlTypeMode.*
 
 /** Abstract representation of the type of a generic SQL database */
 enum SqlType {