+++ spark/src/test/scala/org/datatools/bigdatatypes/spark/SqlTypeToSparkImplicitFormatsSpec.scala
@@ -1,8 +1,8 @@
 package org.datatools.bigdatatypes.spark
 
-import org.apache.spark.sql.types._
+import org.apache.spark.sql.types.*
 import org.datatools.bigdatatypes.formats.{Formats, SnakifyFormats}
-import org.datatools.bigdatatypes.spark.SqlTypeToSpark._
+import org.datatools.bigdatatypes.spark.SqlTypeToSpark.*
 import org.datatools.bigdatatypes.UnitSpec
 
 /** Testing implicit formats