+++ spark/src/test/scala/org/datatools/bigdatatypes/spark/SqlTypeToSparkMultipleClassesSpec.scala
@@ -1,8 +1,8 @@
 package org.datatools.bigdatatypes.spark
 
-import org.apache.spark.sql.types._
+import org.apache.spark.sql.types.*
 import org.datatools.bigdatatypes.formats.{DefaultFormats, Formats}
-import org.datatools.bigdatatypes.spark.SqlTypeToSpark._
+import org.datatools.bigdatatypes.spark.SqlTypeToSpark.*
 import org.datatools.bigdatatypes.UnitSpec
 
 /** Testing multiple case classes for conversions
@@ -76,4 +76,3 @@ class SqlTypeToSparkMultipleClassesSpec extends UnitSpec {
   }
 
 }
-